+++ R/read_xlsx.R
@@ -1207,7 +1207,6 @@ read_xlsx <- function(path = NULL) {
 
   obj <- structure(
     list(package_dir = package_dir),
-    .Names = c("package_dir"),
     class = "rxlsx"
   )